rm duplicated jsonu256
